Fix failing PHP 8.5 CI from automated SDK sync v1.0.2
- Fix enum case naming (UPPERCASE → TitleCase): AgentDiscoveryPathScope,
  SkillDiscoveryScope, InstructionDiscoveryPathKind,
  SubagentSettingsEntryContextTier, McpServerConfigDeferTools
- Fix ModelBillingTokenPrices/LongContext: inputPrice/outputPrice as
  float (matching official SDK), not int
- Fix NewV102EnumsTest: correct enum case names and backing values
- Fix NewV102RpcTest: wrong parameter names (scope→excludeHostAgents,
  location→excludeHostInstructions, model→modelId)
- Fix NewV102TypesTest: rewrite all sections to match actual class shapes
- Fix ModelBillingTest: inputMTokenPrice→inputPrice, outputMTokenPrice→outputPrice
- Fix SessionConfigTest: remove non-existent maxEntries from MemoryConfiguration
- Fix docs/jp/session-config.md: remove non-existent maxEntries parameter

+++ src/Types/Rpc/ModelBillingTokenPrices.php
@@ -12,17 +12,17 @@ use Illuminate\Contracts\Support\Arrayable;
 readonly class ModelBillingTokenPrices implements Arrayable
 {
     /**
-     * @param  ?int  $inputPrice  AI Credits cost per billing batch of input tokens
-     * @param  ?int  $outputPrice  AI Credits cost per billing batch of output tokens
-     * @param  ?int  $cachePrice  AI Credits cost per billing batch of cached tokens
+     * @param  ?float  $inputPrice  AI Credits cost per billing batch of input tokens
+     * @param  ?float  $outputPrice  AI Credits cost per billing batch of output tokens
+     * @param  ?float  $cachePrice  AI Credits cost per billing batch of cached tokens
      * @param  ?int  $batchSize  Number of tokens per standard billing batch
      * @param  ?int  $contextMax  Prompt token budget for the default tier
      * @param  ModelBillingTokenPricesLongContext|null  $longContext  Long context tier pricing
      */
     public function __construct(
-        public ?int $inputPrice = null,
-        public ?int $outputPrice = null,
-        public ?int $cachePrice = null,
+        public ?float $inputPrice = null,
+        public ?float $outputPrice = null,
+        public ?float $cachePrice = null,
         public ?int $batchSize = null,
         public ?int $contextMax = null,
         public ?ModelBillingTokenPricesLongContext $longContext = null,
@@ -31,9 +31,9 @@ readonly class ModelBillingTokenPrices implements Arrayable
     public static function fromArray(array $data): self
     {
         return new self(
-            inputPrice: isset($data['inputPrice']) ? (int) $data['inputPrice'] : null,
-            outputPrice: isset($data['outputPrice']) ? (int) $data['outputPrice'] : null,
-            cachePrice: isset($data['cachePrice']) ? (int) $data['cachePrice'] : null,
+            inputPrice: isset($data['inputPrice']) ? (float) $data['inputPrice'] : null,
+            outputPrice: isset($data['outputPrice']) ? (float) $data['outputPrice'] : null,
+            cachePrice: isset($data['cachePrice']) ? (float) $data['cachePrice'] : null,
             batchSize: isset($data['batchSize']) ? (int) $data['batchSize'] : null,
             contextMax: isset($data['contextMax']) ? (int) $data['contextMax'] : null,
             longContext: isset($data['longContext']) ? ModelBillingTokenPricesLongContext::fromArray($data['longContext']) : null,
